Adapt tests
