+++ tests/neg-macros/i4947e/Macro_1.scala
