+++ tests/neg-macros/i4947e/Test_2.scala
@@ -4,7 +4,7 @@ object Test {
 
   def main(args: Array[String]): Unit = {
     val x = 1
-    assert2(x != 0)
+    assert2(x != 0) // error
     assert2(x == 0)
     assert2 {
       Macros.printStack("hi")
