+++ tests/neg-macros/quote-sep-comp-2/Macro_1.scala
