+++ tests/run-macros/inline-case-objects/Macro_1.scala
@@ -3,6 +3,7 @@ import scala.quoted._
 import scala.quoted.autolift._
 
 object Macros {
+  inline def fooString(inline x: Any): String = ${Macros.impl(x)}
   def impl(foo: Any): Expr[String] = foo.getClass.getCanonicalName
 }
