+++ tests/run-macros/inline-case-objects/Main_2.scala
@@ -1,6 +1,8 @@
 
 object Test {
 
+  import Macros._
+
   def main(args: Array[String]): Unit = {
     println(fooString(scala.collection.immutable.Nil))
     println(fooString(None))
@@ -9,7 +11,5 @@ object Test {
     println(fooString(foo.Bar))
     println(fooString(foo.Bar.Baz))
   }
-
-  inline def fooString(inline x: Any): String = ${Macros.impl(x)}
 
 }
