+++ tests/run-macros/inline-tuples-2/Main_2.scala
@@ -1,5 +1,6 @@
 
 object Test {
+  import Macros._
 
   def main(args: Array[String]): Unit = {
     println(get1(Tuple1(1)))
@@ -10,13 +11,5 @@ object Test {
 
     println(get4(Tuple1(Tuple1(6))))
   }
-
-  inline def get1(inline tup: Tuple1[Int]): Int = ${ Macros.impl(tup) }
-
-  inline def get2(inline i: Int): Int = ${ Macros.impl(Tuple1(i)) }
-
-  inline def get3(inline i: Int): Int = ${ Macros.impl2(Tuple1(Tuple1(i))) }
-
-  inline def get4(inline tup: Tuple1[Tuple1[Int]]): Int = ${ Macros.impl2(tup) }
 
 }
