+++ tests/run-macros/inline-varargs-1/Main_2.scala
@@ -2,8 +2,7 @@
 object Test {
 
   def main(args: Array[String]): Unit = {
-    println(sum(1, 2, 3))
+    println(Macros.sum(1, 2, 3))
   }
 
-  inline def sum(inline i: Int, inline j: Int, inline k: Int): Int = ${ Macros.sum(i, j, k) }
 }
